refactor: create separate file for regular optimizer test
